+++ src/views/_mail.php
@@ -1,25 +1,21 @@
 <?php
+use yii\widgets\DetailView;
 /* @var $model \yii\base\Model The model which holds the migration*/
 /* @var $title string The title text */
 /* @var $text string The message text */
 ?>
 <h2><?= $title; ?></h2>
+<p><i><?= strftime('%c')?></i></p>
 <?= $text; ?>
-<table border="0" cellpadding="5" cellspacing="2" width="100%">
-    <?php foreach ($model->getAttributes() as $key => $value): ?>
-    	<tr>
-    		<td width="150" style="border-bottom:1px solid #F0F0F0"><?= $model->getAttributeLabel($key); ?>:</td>
-    		<td style="border-bottom:1px solid #F0F0F0">
-    			<?php if (is_array($value)): ?>
-    				<ul>
-    					<?php foreach ($value as $item): ?>
-    					<li><?= nl2br($item); ?></li>
-    					<?php endforeach; ?>
-    				</ul>
-    			<?php else: ?>
-    				<?= nl2br($value); ?>
-    			<?php endif; ?>
-    		</td>
-    <?php endforeach; ?>
-</table>
-<p><i><?= date("r"); ?></i></p>
+<?= DetailView::widget([
+    'model' => $model,
+    'attributes' => $detailViewAttributes,
+    'options' => [
+        'class' => null,
+        'style' => 'width:100%',
+        'cellpadding' => 5,
+        'cellpsacing' => 2,
+        'border' => 0,
+    ],
+    'template' => '<tr><th width="150" style="border-bottom:1px solid #F0F0F0" {captionOptions}>{label}</th><td style="border-bottom:1px solid #F0F0F0" {contentOptions}>{value}</td></tr>'
+]); ?>