+++ CHANGELOG.md
@@ -3,6 +3,12 @@ LUYA MODULE CONTACTFORM CHANGELOG
 
 All notable changes to this project will be documented in this file. This project make usage of the [Yii Versioning Strategy](https://github.com/yiisoft/yii2/blob/master/docs/internals/versions.md).
 
+1.0.2 (4. May 2017)
+------------------
+
++ [#4](https://github.com/luyadev/luya-module-contactform/issues/4) Used yii\widgets\DetailView in order to render E-Mails.
++ Added basic Unit-Tests.
+
 1.0.1 (3. May 2017)
 -------------------